refactor: ParameterSliderのインポートパスをv2に変更
