+++ src/components/Talk/v2/ParameterSlider.vue
